requirements.txt
streamlit
openai
python-docx
pydub
ffmpeg-python
soundfile
librosa